First public release build

+++ lib/analyzer_module.ex
@@ -1,12 +1,13 @@
-# Copyright (C) 2018 by the Georgia Tech Research Institute (GTRI)
+# Copyright (C) 2020 by the Georgia Tech Research Institute (GTRI)
 # This software may be modified and distributed under the terms of
 # the BSD 3-Clause license. See the LICENSE file for details.
 
 defmodule AnalyzerModule do
-  
   @moduledoc """
-  Analyzer takes in a repo url and coordinates the analysis,
-  returning a simple JSON report.
+  Analyzer takes in a valid repo URL and coordinates the analysis,
+  returning a simple JSON report.  The URL can be one of "https", "http",
+  or "file".  Note, that the latter scheme will only work an existing clone
+  and won't remove the directory structure upon completion of analysis.
   """
   @spec analyze(String.t() | list(), String.t()) :: tuple()
   def analyze(url, source) when is_binary(url) do
@@ -22,12 +23,18 @@ defmodule AnalyzerModule do
       #  load/start it or list it under :extra_applications in your mix.exs file"
       _config = Application.fetch_env!(:lowendinsight, :critical_contributor_level)
 
+      uri = URI.parse(url)
 
-      if Helpers.count_forward_slashes(url) > 4 do
-        raise ArgumentError, message: "Not a Git repo URL, is a subdirectory"
+      {:ok, repo} = cond do
+        uri.scheme == "file" ->
+          GitModule.get_repo(uri.path)
+        uri.scheme == "https" or uri.scheme == "http" ->
+          if Helpers.count_forward_slashes(url) > 4 do
+            raise ArgumentError, message: "Not a Git repo URL, is a subdirectory"
+          end
+          {:ok, tmp_path} = Temp.path "lei"
+          GitModule.clone_repo(url, tmp_path)
       end
-
-      {:ok, repo} = GitModule.clone_repo(url)
 
       # Get unique contributors count
       {:ok, count} = GitModule.get_contributor_count(repo)
@@ -57,15 +64,22 @@ defmodule AnalyzerModule do
 
       {:ok, top10_contributors} = GitModule.get_top10_contributors_map(repo)
 
-      GitModule.delete_repo(repo)
+      project_types = ProjectIdent.project_types?(repo)
 
-      # Generate report
+      if uri.scheme == "https" or uri.scheme == "http" do
+        GitModule.delete_repo(repo)
+      end
+
       end_time = DateTime.utc_now()
       duration = DateTime.diff(end_time, start_time)
 
       # Return summary report as JSON
       # Workaround to allow `mix analyze` to work even that :application doesn't exist
-      library_version = if :application.get_application != :undefined, do: elem(:application.get_key(:lowendinsight, :vsn), 1) |> List.to_string, else: ""
+      library_version =
+        if :application.get_application() != :undefined,
+          do: elem(:application.get_key(:lowendinsight, :vsn), 1) |> List.to_string(),
+          else: ""
+
       report = %{
         header: %{
           start_time: DateTime.to_iso8601(start_time),
@@ -78,6 +92,7 @@ defmodule AnalyzerModule do
         data: %{
           config: Helpers.convert_config_to_list(Application.get_all_env(:lowendinsight)),
           repo: url,
+          project_types: Helpers.convert_config_to_list(project_types),
           results: %{
             contributor_count: count,
             contributor_risk: count_risk,
@@ -92,12 +107,32 @@ defmodule AnalyzerModule do
           }
         }
       }
+
       {:ok, determine_toplevel_risk(report)}
     rescue
       MatchError ->
-        {:ok, %{data: %{config: Helpers.convert_config_to_list(Application.get_all_env(:lowendinsight)), error: "Unable to analyze the repo (#{url}), is this a valid Git repo URL?", repo: url, risk: "critical"}}}
+        {:ok,
+         %{
+           data: %{
+             config: Helpers.convert_config_to_list(Application.get_all_env(:lowendinsight)),
+             error: "Unable to analyze the repo (#{url}), is this a valid Git repo URL?",
+             repo: url,
+             risk: "critical",
+             project_types: %{"undetermined" => "undetermined"}
+           }
+         }}
+
       e in ArgumentError ->
-        {:ok, %{data: %{config: Helpers.convert_config_to_list(Application.get_all_env(:lowendinsight)), error: "Unable to analyze the repo (#{url}). #{e.message}", repo: url, risk: "N/A"}}}
+        {:ok,
+         %{
+           data: %{
+             config: Helpers.convert_config_to_list(Application.get_all_env(:lowendinsight)),
+             error: "Unable to analyze the repo (#{url}). #{e.message}",
+             repo: url,
+             risk: "N/A",
+             project_types: %{"undetermined" => "undetermined"}
+           }
+         }}
     end
   end
 
@@ -115,26 +150,38 @@ defmodule AnalyzerModule do
     2
     ```
   """
-  #@defaults %{start_time: DateTime.utc_now()}
+  # @defaults %{start_time: DateTime.utc_now()}
   def analyze(urls, source, start_time \\ DateTime.utc_now()) when is_list(urls) do
-    #%{start_time: start_time} = Enum.into(opts, @defaults)
+    # %{start_time: start_time} = Enum.into(opts, @defaults)
     ## Concurrency for parallelizing the analysis.  Turn the analyze/2 function into a worker.
-    l = urls
-      |> Task.async_stream(__MODULE__, :analyze, [source], [timeout: :infinity, max_concurrency: 10])
+    l =
+      urls
+      |> Task.async_stream(__MODULE__, :analyze, [source], timeout: :infinity, max_concurrency: 10)
       |> Enum.map(fn {:ok, report} -> elem(report, 1) end)
-    report = %{state: "complete", report: %{uuid: UUID.uuid1(), repos: l}, metadata: %{repo_count: length(l)}}
+
+    report = %{
+      state: "complete",
+      report: %{uuid: UUID.uuid1(), repos: l},
+      metadata: %{repo_count: length(l)}
+    }
 
     report = determine_risk_counts(report)
     end_time = DateTime.utc_now()
     duration = DateTime.diff(end_time, start_time)
-    times = %{start_time: DateTime.to_iso8601(start_time), end_time: DateTime.to_iso8601(end_time), duration: duration}
+
+    times = %{
+      start_time: DateTime.to_iso8601(start_time),
+      end_time: DateTime.to_iso8601(end_time),
+      duration: duration
+    }
+
     metadata = Map.put_new(report[:metadata], :times, times)
     report = report |> Map.put(:metadata, metadata)
     {:ok, report}
   end
 
   @doc """
-  create_empty_report/3: takes in a uuid, list of urls, and a start time and 
+  create_empty_report/3: takes in a uuid, list of urls, and a start time and
   produces the repo report object to be returned immediately by asynchronous
   requestors (e.g. LowEndInsight-Get HTTP endpoint)
   """
@@ -163,9 +210,11 @@ defmodule AnalyzerModule do
   format - so caching can be supported (as reports are stored in JSON).
   """
   def determine_risk_counts(report) do
-    count_map = report[:report][:repos]
-        |> Enum.map(fn (repo) -> repo[:data][:risk] end)
-        |> Enum.reduce(%{}, fn x, acc -> Map.update(acc, x, 1, &(&1 + 1)) end)
+    count_map =
+      report[:report][:repos]
+      |> Enum.map(fn repo -> repo[:data][:risk] end)
+      |> Enum.reduce(%{}, fn x, acc -> Map.update(acc, x, 1, &(&1 + 1)) end)
+
     metadata = Map.put_new(report[:metadata], :risk_counts, count_map)
     report |> Map.put(:metadata, metadata)
   end
